API Version 0.1

+++ src/main/java/pl/proteingeekbackend/model/Product.java
@@ -2,7 +2,7 @@ package pl.proteingeekbackend.model;
 
 public class Product {
 
-    private int id;
+    private double id;
     private String name;
     private String brand;
     private int Volume;
